Face_ID_and_Emote/requirements.txt
facenet_pytorch==2.6.0
opencv-python==4.9.0
torch==2.2.0+cu121
pandas==2.0.3
numpy==1.26.4
pypyodbc==1.3.6
Pillow==10.2.0
torchvision==0.17.0+cu121